Add files via upload
The primary goal of a diabetic retinopathy project is to develop an automated system capable of detecting and classifying DR based on retinal images, which are often taken via specialized cameras known as fundus cameras. These cameras capture detailed images of the retina, and trained medical professionals typically examine them for signs of DR. However, manual diagnosis is time-consuming and requires significant expertise. With the help of technology, we can build a system that can quickly and accurately detect diabetic retinopathy, making it accessible to a larger population and reducing the workload on healthcare providers.